writing-kana/requirements.txt
streamlit==1.32.0
pillow==10.2.0
numpy==1.26.3
manga-ocr==0.1.10